Refactor Exceptions
